Add initial solution for calculating average grade
This initial solution addresses the task of calculating the average grade for a given number of students. The code prompts the user for the number of students and their corresponding grades. It then computes the average grade by summing up all the grades and dividing by the number of students. The result is displayed with two decimal places for precision.