Add solution for rotating elements in a list
This commit introduces a solution for rotating elements within a list. The code reads an input list of integers and the number of rotations to perform. It then repeatedly moves the last element to the front of the list for the specified number of rotations. Finally, it displays the modified list.